[wip] repo: Add RFC process document
This process is adapted from the Rust RFC process on rust-lang/rfcs and
is intentionally less formal than that of Rust as the Typst community
is still much smaller and the RFCs are less official than those of Rust.

The formality may change in the future and the repository may be
incorporated into the typst organization itself. But until then, this
process can stay simple and somewhat informal.